Move PD controller out to the utilities.
