+++ examples/cylinder/pressure-probes.py
@@ -2,7 +2,6 @@
 Simulate the flow around the cylinder with evenly spaced surface pressure probes.
 """
 import os
-
 import matplotlib.pyplot as plt
 import numpy as np
 import psutil
